eatgo-admin-api/src/main/java/me/hangyeol/eatgo/region/service/RegionService.java
package me.hangyeol.eatgo.region.service;

import me.hangyeol.eatgo.region.Region;
import me.hangyeol.eatgo.region.RegionRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class RegionService {

    private RegionRepository regionRepository;

    public RegionService(RegionRepository regionRepository) {
        this.regionRepository = regionRepository;
    }

    public List<Region> getRegions() {
        return regionRepository.findAll();
    }

    public Region addRegion(String name) {
        Region region = Region.builder().name(name).build();
        regionRepository.save(region);
        return region;
    }
}
